Add support for `docker volume ls`
When the driver attempts a `List` in `plugin.go`,
it calls `vmdkops.VmdkList()`, which now returns a pair of type `([]volume.Volume, error)`.
If there was no error then the slice returned is used to build up a client response.
In order to retrieve this list of Volumes, Json plumbing was required at the go/C boundary
in `vmdkops.go` and `vmci_client.c`. This plumbing was also required of The python server
running on ESX in `vmci_srv.py`.

In order to prevent having to change the entire API at once, backwards compatibility was
maintained with all other functions beside List. These should all be updated in another PR.