Fix docker volume ls
When using a range loop in go you can't take a pointer to the value of the object returned.
It turns out that this is a temporary variable where the value is copied, so everytime you
 end up with the same pointer. Instead of using the value, you have to use the index and
then index the original array. This implicit pointer nonsense made for some fun debugging.