Return errors from ESX service in json
When an error occurred in the ESX service, it was of a tuple type. This
failed to serialize as json and caused the ESX service to crash. This
change now properly encodes the errors as JSON so they can be returned
to the plugin on the guest, and then to the user at the docker
cli.

As part of this change RunCommand in vmci_srv.py no longer raises
exceptions. Instead it always returns a two tuple of return value and
either stdout or stderr depending upon whether there was an error or
not.

Also add unlink and formatVMDK functions in vmci_srv.py to enhance
readability