Converted tests from Makefile to GO test
This is the first phase of testing code development. It does not add new
tests - just different code and more info in logs.
The next phase will be extending test set.

Also, did some cosmetic changes to testremote-wrapper as the amount of stars
it used to print made if very hard to find log sections