Moved deploy and cleanup logic from Makefiles (issue 180)
To properly handle multiple guests (and ESXs in the future) for deployment
and cleanup, as well as to provide some flexibility in changing this code and not
be limited to Makefiles, I moved the logic out. It should work for both 'make' and drone CI
as the latter works via make.

This also should address Issue 180