First working version of python unittests. Somewhat trivial
Allows to run unittests on ESX. Does not require running vmdkops service, but does
current does expect mkfs.ext4 in the proper location.

To run:

make deploy-esx # this is only needed to place mkfs.ext4
make test-esx   # copies and runs unittests, and cleans them up

leaves log in  /var/log/vmware/docker-vmdk-plugin-pytest.log.
Note that the log is not truncated and and not roatated (until
we fix log rotation in  python service)