Discovering volume usage refcounts from Docker on plugin-in start
For each container using a volume, Docker send us a Mount on start and Unmount on stop.
So if multiple containers using a volume, we need our own refcounters for volume usage.

This commit separates existing in-memory refcounters into a dedicated file and reworks the code
to support discovery-on-start and basic recovery:
When the plugin is restarted, we discover the refcounters from Docker container info.

1. on start, plugin connects to Docker, enumerates all container Mounts, and
builds refcount map.
2. /proc/mount table is analized and volumes which are not used by docker released.

The code is tested with ./scripts/refcnt_test.sh, which is called in 'go test' as a shell command.

We still need to add crash testing in a separate PR.