Initial version of ESX admin script
This commit contains parsers and tests for ls, df, and policy commands.
These parsers and tests still need to be written for role and status.

No functionality is currently implemented, but the function stubs exist
that get called with the parsed arguments when a command is parsed.

I also added vmdkops_admin.py to the VIB. The test is currently not run
anywhere, but as it only relies on argparse it can be run anywhere right
now.