Add created and created-by fields to admin ls -l
The fields are now stored in sidecar kv. There was some small cleanup of code
along with this. Cleanup included splitting metadata gathering into small
functions in vmdkops_admin.py and creating docstrings in volume_kv.py. Loggins
was removed from volume_kv.py as it wasn't featureful, and users of the library
should be the ones doing the logging.