Made refcnt dicovery conditional on docker connection (#372)
* Made refcnt dicovery conditional on docker connection

If Docker is running (e.g. started before the plugin) , then we discover refcounts
and do recovery mount/unmpunt if needed. This handles Plugin crash use case
Note: During discovery, Mount and Unmount requests are denied to make behavior  predictable

If Docker is not running, simply sckp the dicsovery - Docker is supposed to communicate
proper mounts so we reconstruct refcnts as we go

* Removed initInProgress as we call Init() before starting to accept (un)mount requests