Only process requests from clients with privilege
To be able to bind a port below 1024 on the client side, the plugin
needs to be run as root or given neccessary privileges. This change
enforces the policy of only allowing privileged clients.

To keep things simple the client port to bind is the same as the server.

Testing:
1. Install client without code change to use priviledge port
2. Validate that docker requests fail and log messages are generated.
3. Upgrade to client with code change for priviledge port
4. Validate that docker requests pass

Will be filling a separate issue to resolve the std out logging
on ESX. We need to invest a bit more to override the std out buffer
and redirect it to a log file.