Fix vsan policy test failures on CI
Tests were failing with "out of resources" messages due to the fact that we were
using a `(("hostFailuresToTolerate" i1))` for the original config. This works
when there are actually multiple ESXs in the VSAN cluster, but fails when a
single ESX is used because you can't tolerate one host failuer with one host.

This change uses `(("hostFailuresToTolerate" i0))` for both the old and new
configs and adds `(("proportionalCapacity" i0))` to the original config which is
harmless, but allows for 2 different configs to allow for testing policy update
code paths.

This was tested via `make all` against both our internal testbed with 3 ESX VSAN
cluster and CI with one ESX.