Allow dash in volume names (#567)
* Allowed dash in volume names

And generally relaxed names to have anything as long as there is no conflict with deltadisk names.
Still prevent conflict with deltadisks by blocking
'-NNNNNN' at the end of the volume name. E.g. volume-mySuperHero-2 is a valid
volume name, but volume-Here-000012 is not.
Also removed abuse of regexps (where a simple rindex() is enough) and tidied up volume/vmdk name parse code.

* Fixes to address review comments