Link VIB version to the build version.
The vib version was not being set for each build or release.
As per documentation an additional release version is added.

```
Versioning VIBs

Because VIBs are intended to allow the ESXi installer to handle updates as well as initial installs, it is
important to use the versioning variables in the VIB correctly.

The package version can have a length of up to 35 characters. It includes the version number, immediately
followed by a hyphen (‐) followed by the release version, as in the following example:

5.0.0-1.0vmw

Both the version number and the release version number consist of one or more characters in the set
[a‐a‐Z0‐9], with periods separating two or more groups.
```

Testing:

Make vib and install.

Run command ```esxcli software profile get```

Output:
```
(Original Vendor):VMware, Inc.
      2016-07-06T10:45:12.595760+00:00: The following VIBs are
      installed:
        esx-vmdkops-service   0.7.14b2727-0.0.1
```