Moved admin 'ls' and 'set' commands under 'volume' (#1055)
Rename 'vmdkops_admin ls" to ‘vmdkops_admin volume ls’
and rename ‘vmdkops_admin set’ to ‘ vmdkops_admin volume set’

This is a part of #1043

As I work through :multi-ESX tenancy code, some pieces ask for fix or implementation and make it easier for me to proceed. I do not want to bundle them all in a large PR, so I am submitting individual PRs which are useful on their own. This is one of them,

Test: make test-esx locally. Manual validation. CI