vShared: remove orphan services and internal volumes (#1774)
Due to unexpected errors, the docker services and internal volumes
may not be removed completely when the shared volume is deleted. The
existing of those orphan services and internal volumes will result in
a waste of resources and can hurt the docker performance. It can also
result in error when creating volumes with the previous existed names.

This change adds handling for those orphan services and orphan internal
volumes. We first query all the shared volumes and their states from kv
store. Then we compare if there is any service or internal volume doesn't
have a matching shared volume or state. No matching shared volume
means the service or internal volume is orphan, which need to be removed.