vFile: move time-consuming initialization works to goroutine
When a plugin is started/restarted, docker has a small timeout value.
If the plugin init function is not able to return within this time,
it will fail to enable the plugin. For our vFile plugin, we need a
bunch of extra works such as loading server images and start/join
kvstore. This patch moves those time-consuming initialization tasks
into goroutine, in order to make sure the plugin can be enabled in time.