[SKIP CI] Enables sticky slaves for Jenkins builds (#2069)
This commit updates the logic used for selecting slaves for executing
builds on Jenkins. Every slave (node) is marked with an `available`
label, which the script looks for in the `Slave selection` stage. Once
the nodes are identified, the `available` label is removed an a new UUID
label is applied. For the rest of the build execution, nodes are always
referenced using the UUID label. At the end of the build, these labels
are reverted to make nodes available for other PRs.